modified:   Makefile 	new file:   ep1_lista_13781883.exe 	new file:   ep1_matriz_13781883.exe 	modified:   grafo_listaadj.c 	modified:   grafo_matrizadj.c 	modified:   grafo_matrizadj.h 	modified:   main.c

+++ grafo_matrizadj.h
@@ -4,7 +4,7 @@
 #define AN -1 //ausencia de aresta
 #define VERTICE_INVALIDO -1 //vertice inexistente
 
-typedef int Peso;
+typedef float Peso;
 typedef struct {
     Peso mat[MAXNUMVERTICES + 1][MAXNUMVERTICES + 1];
     int nroVertices;
